Requests for merge
- now using com.dtteam group ID
- moved neoforge.mods.toml
- adjusted species jsons
- used dt addon for magic logic, cell kit, and species type
- ran runData
- removed unused mixin files
- pack format changed
- issue tracker URL added
- adjusted author/credit fields correctly
