+++ src/main/java/com/dtteam/dtvampirism/cell/DTVampirismLeafClusters.java
@@ -1,4 +1,4 @@
-package net.altias.dtvampirism.cell;
+package com.dtteam.dtvampirism.cell;
 
 import com.dtteam.dynamictrees.api.voxmap.SimpleVoxmap;
 import net.minecraft.core.BlockPos;
