--- src/main/java/net/altias/dtvampirism/cell/MagicBranchCell.java
@@ -1,20 +1,0 @@
-package net.altias.dtvampirism.cell;
-
-import com.dtteam.dynamictrees.api.cell.Cell;
-import net.minecraft.core.Direction;
-
-public class MagicBranchCell implements Cell {
-
-    @Override
-    public int getValue() {
-        return 5;
-    }
-
-    static final int[] map = {3, 4, 4, 4, 4, 4};
-
-    @Override
-    public int getValueFromSide(Direction side) {
-        return map[side.ordinal()];
-    }
-
-}
