--- src/main/java/net/altias/dtvampirism/cell/MagicLeafCell.java
@@ -1,21 +1,0 @@
-package net.altias.dtvampirism.cell;
-
-
-import com.dtteam.dynamictrees.systems.cell.MatrixCell;
-
-public class MagicLeafCell extends MatrixCell {
-
-    public MagicLeafCell(int value) {
-        super(value, valMap);
-    }
-
-    static final byte[] valMap = {
-            0, 0, 0, 0, 0, 0, 0, 0, //D Maps * -> 0
-            0, 1, 2, 3, 4, 0, 0, 0, //U Maps 0 -> 0, * -> *
-            0, 1, 1, 1, 1, 0, 0, 0, //N Maps 0 -> 0, * -> 1
-            0, 1, 1, 1, 1, 0, 0, 0, //S Maps 0 -> 0, * -> 1
-            0, 1, 1, 1, 1, 0, 0, 0, //W Maps 0 -> 0, * -> 1
-            0, 1, 1, 1, 1, 0, 0, 0  //E Maps 0 -> 0, * -> 1
-    };
-
-}
