--- src/main/java/net/altias/dtvampirism/cell/MagicTopBranchCell.java
@@ -1,20 +1,0 @@
-package net.altias.dtvampirism.cell;
-
-import com.dtteam.dynamictrees.api.cell.Cell;
-import net.minecraft.core.Direction;
-
-public class MagicTopBranchCell implements Cell {
-
-    @Override
-    public int getValue() {
-        return 5;
-    }
-
-    static final int[] map = {3, 4, 3, 3, 3, 3};
-
-    @Override
-    public int getValueFromSide(Direction side) {
-        return map[side.ordinal()];
-    }
-
-}
