--- src/main/resources/dtvampirism.mixins.json
@@ -1,16 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.altias.dtvampirism.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "mixins": [
-  ],
-  "client": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "overwrites": {
-    "requireAnnotations": true
-  }
-}
