+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
         "description": "Dynamic Trees Vampirism Resources",
-        "pack_format": 15
+        "pack_format": 34
     }
 }
