+++ src/main/resources/trees/dtvampirism/species/dark_spruce.json
@@ -6,12 +6,9 @@
   "lowest_branch_height": 3,
   "growth_rate": 0.9,
   "growth_logic_kit": "conifer",
-  "environment_factors" : {
-    "#forge:is_hot": 0.50,
-    "#forge:is_dry": 0.25,
-    "#forge:is_wet": 0.75
-  },
   "perfect_biomes": { "name": "vampirism:vampire_forest.*" },
   "primitive_sapling": "vampirism:dark_spruce_sapling",
-  "acceptable_soils": [ "dirt_like", "corrupt_dirt_like" ]
+  "acceptable_soils": [ "dirt_like", "corrupt_dirt_like" ],
+  "generate_seed": true,
+  "generate_sapling": true
 }
