+++ src/main/resources/trees/dtvampirism/species/jacaranda.json
@@ -5,10 +5,8 @@
   "signal_energy": 12.0,
   "growth_rate": 0.9,
   "leaves_properties": "dtvampirism:jacaranda",
-  "environment_factors" : {
-    "hot": 0.5,
-    "dry": 0.5
-  },
+  "generate_seed": true,
+  "generate_sapling": true,
   "perfect_biomes": { "name": "werewolves:werewolf_heaven" },
   "primitive_sapling": "werewolves:jacaranda_sapling"
 }
