+++ src/main/resources/trees/dtvampirism/species/magic.json
@@ -1,6 +1,6 @@
 {
   "only_if_loaded": "werewolves",
-  "type": "dtvampirism:magic",
+  "type": "dtaddon_lib:poplar",
   "family": "dtvampirism:magic",
   "generate_seed": true,
   "generate_sapling": true,
@@ -9,7 +9,7 @@
   "up_probability": 5,
   "lowest_branch_height": 4,
   "growth_rate": 0.85,
-  "growth_logic_kit": "dtvampirism:magic",
+  "growth_logic_kit": "dtaddon_lib:poplar",
   "leaves_properties": "dtvampirism:magic",
   "perfect_biomes": { "name": "werewolves:werewolf_heaven" },
   "primitive_sapling": "werewolves:magic_sapling"
